backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
cryptography==41.0.7
Pillow==10.1.0
qrcode[pil]==7.4.2
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pynacl== 1.6.2
pyzbar==0.1.9
numpy==1.26.4
easyocr==1.7.1
torch>=2.0.0
torchvision>=0.15.0
python-jose[cryptography]
passlib[bcrypt]
jwt-decode